+++ src/Schema/AbstractSqlTools.php
@@ -2,7 +2,7 @@
 /**
  * @author     Gérald Croes, Laurent Jouanneau
  * @contributor Laurent Jouanneau, Julien Issler
- * @copyright  2001-2005 CopixTeam, 2005-2023 Laurent Jouanneau, 2008 Julien Issler
+ * @copyright  2001-2005 CopixTeam, 2005-2025 Laurent Jouanneau, 2008 Julien Issler
  *
  * @see        https://jelix.org
  * @licence     http://www.gnu.org/licenses/lgpl.html GNU Lesser General Public Licence, see LICENCE file
@@ -61,6 +61,7 @@ abstract class AbstractSqlTools implements SqlToolsInterface
         'blob' => 'string',
         'binary' => 'string',
         'varbinary' => 'string',
+        'json' => 'array',
     );
 
     protected $typesInfo = array();
@@ -199,6 +200,10 @@ abstract class AbstractSqlTools implements SqlToolsInterface
             case 'numeric':
             case 'decimal':
                return Utilities::floatToStr($value);
+            case 'array':
+                if (!is_string($value)) {
+                    $value = json_encode($value);
+                }
             default:
                 if ($toPhpSource) {
                     if ($unifiedType == 'varbinary' || $unifiedType == 'binary') {
@@ -660,6 +665,10 @@ abstract class AbstractSqlTools implements SqlToolsInterface
                         $val = 'NULL';
                         $op = 'IS';
 
+                        break;
+                    case 'array':
+                    case 'object':
+                        $val = $this->_conn->quote(json_encode($value));
                         break;
                     default:
                         $this->_conn->rollback();